Move 'property' to its own file
